# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/analysis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Get analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization header with Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "500": {
                        "description": "An error occurred, please try again later",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            }
        },
        "/analysis/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get analysis by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Get analysis by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header with Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid analyze ID",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "500": {
                        "description": "An error occurred, please try again later",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Analyze with Suricata",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Analyze with Suricata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid analyze ID",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Delete analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid analyze ID",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "500": {
                        "description": "An error occurred, please try again later",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            }
        },
        "/analysis/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download analysis",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Download analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid analyze ID",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            }
        },
        "/analysis/{id}/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get exported files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Get exported files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid analyze ID",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            }
        },
        "/analysis/{id}/files/{file}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download exported file",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Analyzer"
                ],
                "summary": "Download exported file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "file",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid analyze ID",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "$ref": "#/definitions/types.FailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.FailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}